requirements.txt
torch==2.3.0
numpy
torchvision
Pillow
transformers==4.45.2
termcolor
accelerate
datasets==3.0.1
huggingface_hub
tensorboard
wandb
packaging
tqdm
pyyaml
xformers
einops
peft==0.14.0
scipy
safetensors
opencv-python
k-diffusion
omegaconf
pytorch-lightning==1.4.2
open_clip_torch==2.0.2
timm
addict
yapf
prettytable
basicsr
gdown